Refactoring client listener service
Smart and non-smart client listener services do the same thing in high level. They both register listeners to all available connections. Smart one has multiple connections available whereas non-smart has only one(owner). Still, the operation is to register to all available connections. Non-smart listener is a special case of smart listener service. Apart from this, the only difference between the two is that smart listeners register to only local events by setting localOnly flag in registration requests. This is easily abstracted using an abstract method in AbstractClientListenerService.

This PR reduces the code duplication between smart and non-smart client listener services.
